
\name{ACEdata}

\alias{ACEdata}

\docType{data}

\title{ACE data set}

\description{ACE (angiotensin I converting enzyme) data set}

\usage{data(ACEdata)}

\format{A data set with 22 haplotypes and 52 SNPs.}

%\details{}

% \source{ \url{http://htsnp.stanford.edu/PCA/ACE.html} }

\references{
Montana, G. HapSim: a simulation tool for generating haplotype data with
pre-specified allele frequencies and LD coefficients. 2005.
}

% \examples{}

\keyword{datasets}
