\name{checkpd}

\alias{checkpd}


\title{Checks that a covariance matrix is positive definite}

\description{
  Checks that a matrix is positive definite
}

\usage{
checkpd(sigma, tol = 1e-06)
}


\arguments{
  \item{sigma}{Covariance matrix}
  \item{tol}{Tolerance value}
}

%\details{}

\value{
If TRUE, the input matrix is positive definite.
}

%\references{}

%\author{}

%\note{}

%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }

%\examples{}

\keyword{ internal }