\name{cor2dprime}

\alias{cor2dprime}


\title{Correlations to D' coefficients conversion}

\description{
  Converts a matrix of correlation coefficients into a D' matrix
}

\usage{
cor2dprime(mat, probs)
}


\arguments{
  \item{mat}{A correlation matrix}
  \item{probs}{A vector of allele frequencies}
}

% \details{}

\value{
 The resulting matrix of D' coefficients
}

\references{
Montana, G. HapSim: a simulation tool for generating haplotype data with
pre-specified allele frequencies and LD coefficients. 2005.
}

\author{Giovanni Montana}

%\note{ ~~further notes~~ }

%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }

% \examples{}

\keyword{internal}
