\name{cortocov}

\alias{cortocov}

\title{Correlation to covariance conversion}

\description{
Converts a correlation matrix into a covariance matrix
}

\usage{
cortocov(cor.mat, sd)
}

\arguments{
  \item{cor.mat}{Correlation matrix}
  \item{sd}{Vector of standard devations}
}

%\details{}

\value{
The covariance matrix
}

%\references{ ~put references to the literature/web site here ~ }

%\author{ ~~who you are~~ }

%\note{ ~~further notes~~ }

%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }

%\examples{}

\keyword{ internal }
