\name{covtocor}

\alias{covtocor}

\title{Covariance to correlation conversion}

\description{
Converts a covariance matrix into a correlation matrix
}

\usage{
covtocor(cov.mat)
}

\arguments{
  \item{cov.mat}{Covariance matrix}
}

%\details{}

\value{
The correlation matrix
}

%\references{ ~put references to the literature/web site here ~ }

%\author{ ~~who you are~~ }

%\note{ ~~further notes~~ }

%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }

%\examples{}

\keyword{ internal }
