\name{divlocus}

\alias{divlocus}

\title{Diversity score}

\description{
Compute an index of genetic diversity at a locus
}

\usage{
divlocus(x)
}

\arguments{
  \item{x}{A vector of binary markers at a locus}
}

%\details{}

\value{
The diversity score
}

%\references{ ~put references to the literature/web site here ~ }

%\author{ ~~who you are~~ }

%\note{ ~~further notes~~ }

%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }

\examples{}

\keyword{internal}