\name{makepd}

\alias{makepd}

\title{Makes a matrix positive definite}

\description{
Modifies a matrix so that it is positive definite
}

\usage{
makepd(mat, eig.tol = 1e-06)
}

\arguments{
  \item{mat}{Matrix}
  \item{eig.tol}{Tolerance value}
}

%\details{}

\value{
A positive-definite matrix
}

%\references{ ~put references to the literature/web site here ~ }

%\author{ ~~who you are~~ }

%\note{ ~~further notes~~ }


%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }

%\examples{}

\keyword{internal}