\name{mse}

\alias{mse}

\title{Mean Square Error}

\description{
Computes the Mean Square Error between two vectors
}

\usage{
mse(x, y)
}

\arguments{
  \item{x}{First vector}
  \item{y}{Second vector, must have same lenght}
}

%\details{}

\value{
The MSE score
}

%\references{ ~put references to the literature/web site here ~ }

%\author{ ~~who you are~~ }

%\note{ ~~further notes~~ }

%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }

%\examples{}

\keyword{internal}