\name{sumsqscale}
\alias{sumsqscale}

\title{ Scales a matrix }

\description{
  Scales the columns of a matrix so that the squared elements sum to unity
}

\usage{
sumsqscale(mat)
}

\arguments{
  \item{mat}{A matrix}
}

%\details{}

\value{
The scaled matrix.
}

%\references{ ~put references to the literature/web site here ~ }

%\author{ ~~who you are~~ }

%\note{ ~~further notes~~ }


%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }

%\examples{}

\keyword{internal}
